refactor :: user 부분 리팩터링
